add createreceipt activity
