+++ app/src/main/res/layout/recycler_item_receipt_detail.xml
@@ -74,12 +74,18 @@
                         android:textColor="#333"
                         android:textStyle="bold" />
 
+
                     <ImageButton
+                        android:background="@drawable/custom_edit_text_border"
                         android:layout_width="24dp"
-                        android:layout_height="24dp"
                         android:id="@+id/btnMinus"
-                        android:background="@color/lightgray"
-                        android:src="@drawable/ic_minus" />
+                        android:layout_height="24dp"
+                        android:src="@drawable/ic_minus"
+                        android:inputType="text"
+                        android:padding="15dp"
+                        android:textColor="@color/black"
+                        android:textColorHint="@color/black"
+                        android:textSize="15sp" />
 
                     <TextView
                         android:id="@+id/txtQuantity"
@@ -91,14 +97,21 @@
                         android:textColor="#333" />
 
                     <ImageButton
+                        android:background="@drawable/custom_edit_text_border"
                         android:layout_width="24dp"
                         android:id="@+id/btnPlus"
                         android:layout_height="24dp"
-                        android:background="@color/lightgray"
-                        android:src="@drawable/ic_add" />
+                        android:src="@drawable/ic_add"
+                        android:inputType="text"
+                        android:padding="15dp"
+                        android:textColor="@color/black"
+                        android:textColorHint="@color/black"
+                        android:textSize="15sp" />
 
 
                 </LinearLayout>
+
+
 
                 <LinearLayout
                     android:layout_width="wrap_content"
